ft_strlen

+++ Libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: anonymous <anonymous@student.42tokyo.jp    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/15 10:40:04 by anonymous         #+#    #+#             */
-/*   Updated: 2023/01/15 11:00:45 by anonymous        ###   ########.fr       */
+/*   Updated: 2023/01/15 11:27:53 by anonymous        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	const char	*ptr;
